MisplaceAI/requirements.txt
Django==3.2
mysqlclient==2.1.0
Pillow==9.3.0
tensorflow
opencv-python==4.5.5.64
djangorestframework==3.12.4
django-cors-headers==3.10.0
djangorestframework-simplejwt==4.6.0
